third_party/blink/web_tests/pointer-lock/pointerlockelement-null-when-pending-expected.txt

Test pointerLockElement is null when a lock is pending.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.pointerLockElement is null
     Lock.
PASS document.pointerLockElement is null
PASS onpointerlockchange received after: Lock.
PASS document.pointerLockElement is targetDiv1
     Unlock.
PASS onpointerlockchange received after: Unlock.
PASS document.pointerLockElement is null
PASS successfullyParsed is true

TEST COMPLETE
